Support git branch in output dir when doing regenerate

+++ firm.cmd
@@ -43,8 +43,26 @@ echo Done!
 goto :end
 
 :regenerate
-echo Removing output...
-rd /s /q output
+if exist output (
+    if exist output\.git (
+        echo Backing up .git and removing output...
+        if exist output-temp (
+            echo Error! The directory output-temp already exists, please remove it manually before retrying regenerate.
+            goto :end
+        )
+        md output-temp\.git
+        xcopy output\.git output-temp\.git /E /H /Q /K
+        rd /s /q output
+        echo Restoring output\.git...
+        md output\.git
+        xcopy output-temp\.git output\.git /E /H /Q /K
+        rd /s /q output-temp
+    ) else (
+        echo Removing output...
+        rd /s /q output
+    )
+    echo Done!
+)
 :generate
 if not exist bin call :build build
 echo Generating site...
